requirements.txt
bcrypt==3.2.0
cffi==1.14.4
Django==2.2.4
django-filter==2.4.0
Pillow==8.0.1
pycparser==2.20
pytz==2020.4
regex==2020.11.13
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
